chore: rename project to `i18n-tools`
yes, I know I'm changing it again, this just fits better with the tools that I find myself needing to build here... including one I'm currently working on to convert back and forth between a flattened JSON file and a DOCX, intended to be sent for translation by a third-party.

this may also be moved to a new `@rhp` NPM scope if I ever decide to publish it to the NPM registry.